[#76] newHomeView -> HomeView로 변경
